+++ src/index.ts
@@ -106,6 +106,7 @@ export type {
     SocialLink,
     TopBar,
     CollectionConfig,
+    CollectionIndexCopy,
     CollectionFields,
     CollectionReference,
     FieldNames,